refactor: remove unneeded consts, move banner fn
Remove unnecessary constant definitions, primarily in the banner
printing function. Also move out the function from the constants module
to its own module `lyra/src/core/banner.rs`.
